Refactor the auth screens and widgets and bloc files a bit
