+++ lib/core/common/widgets/custom_text_field.dart
@@ -1,15 +1,17 @@
 import 'package:flutter/material.dart';
 
-class AuthField extends StatelessWidget {
+class CustomTextField extends StatelessWidget {
   final String hintText;
   final TextEditingController controller;
   final bool isObscureText;
+  final int? maxLines;
 
-  const AuthField({
+  const CustomTextField({
     super.key,
     required this.hintText,
     required this.controller,
     this.isObscureText = false,
+    this.maxLines = 1,
   });
 
   @override
@@ -19,6 +21,7 @@ class AuthField extends StatelessWidget {
       decoration: InputDecoration(
         hintText: hintText,
       ),
+      maxLines: maxLines,
       validator: (value) {
         if(value!.isEmpty) {
           return '$hintText is missing';
