+++ lib/features/auth/presentation/pages/login_page.dart
@@ -1,10 +1,10 @@
-import 'package:blog_app/core/common/widgets/loader.dart';
+import 'package:blog_app/core/common/widgets/custom_text_field.dart';
+import 'package:blog_app/core/common/widgets/custom_loader.dart';
 import 'package:blog_app/core/theme/app_pallete.dart';
 import 'package:blog_app/core/utils/show_snackbar.dart';
 import 'package:blog_app/features/auth/presentation/bloc/auth_bloc.dart';
 import 'package:blog_app/features/auth/presentation/pages/signup_page.dart';
-import 'package:blog_app/features/auth/presentation/widgets/auth_field.dart';
-import 'package:blog_app/features/auth/presentation/widgets/auth_gradient_button.dart';
+import 'package:blog_app/core/common/widgets/custom_gradient_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -59,18 +59,18 @@ class _LoginPageState extends State<LoginPage> {
                     ),
                   ),
                   const SizedBox(height: 30),
-                  AuthField(
+                  CustomTextField(
                     hintText: 'Email',
                     controller: emailController,
                   ),
                   const SizedBox(height: 15),
-                  AuthField(
+                  CustomTextField(
                     hintText: 'Password',
                     controller: passwordController,
                     isObscureText: true,
                   ),
                   const SizedBox(height: 20),
-                  AuthGradientButton(
+                  CustomGradientButton(
                     buttonText: 'Login',
                     onPressed: () {
                       if (formKey.currentState!.validate()) {
