+++ lib/features/blogs/presentation/bloc/blog_bloc.dart
@@ -1,6 +1,5 @@
 import 'dart:io';
 
-import 'package:blog_app/core/error/failures.dart';
 import 'package:blog_app/features/blogs/domain/use_cases/upload_blog.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
